+++ mowl/base_models/elmodel.py
@@ -3,7 +3,9 @@ from mowl.base_models.model import EmbeddingModel
 import torch as th
 from torch.utils.data import DataLoader, default_collate
 from mowl.datasets.el import ELDataset
+from deprecated.sphinx import versionadded
 
+from org.semanticweb.owlapi.model import OWLClassExpression, OWLClass, OWLObjectSomeValuesFrom, OWLObjectIntersectionOf
 
 class EmbeddingELModel(EmbeddingModel):
     """Abstract class that provides basic functionalities for methods that aim to embed EL \
@@ -37,6 +39,10 @@ merging the 3 extra to their corresponding origin normal forms. Defaults to True
         self._training_datasets = None
         self._validation_datasets = None
         self._testing_datasets = None
+
+
+    def init_module(self):
+        raise NotImplementedError
 
     def _load_datasets(self):
         """This method will create different data attributes and finally the corresponding \
@@ -162,3 +168,95 @@ of :class:`torch.utils.data.DataLoader`
 
         self._load_dataloaders()
         return self._testing_dataloaders
+
+    @versionadded(version="0.1.2")
+    def score(self, axiom):
+        def data_point_to_tensor(data_point):
+            data_point = th.tensor(data_point, dtype=th.long, device=self.device)
+            data_point = data_point.unsqueeze(0)
+            return data_point
+        
+        not_el_error_msg = "This axiom does not belong to the EL description logic specification."
+        sub, super_ = axiom.getSubClass(), axiom.getSuperClass()
+
+        if not isinstance(sub, OWLClassExpression):
+            raise TypeError("Parameter sub must be of type OWLClassExpression.")
+
+        if isinstance(sub, OWLClass):
+            sub_id = self.dataset.class_to_id[sub]
+            if isinstance(super_, OWLClass):
+                super_id = self.dataset.class_to_id[super_]
+                if super_.isOWLNothing():
+                    if self.extended:
+                        gci_name = "gci0_bot"
+                    else:
+                        gci_name = "gci0"
+                else:
+                    gci_name = "gci0"
+
+                gci_data = data_point_to_tensor([sub_id, super_id])
+
+            elif isinstance(super_, OWLObjectSomeValuesFrom):
+                rel = super_.getProperty()
+                filler = super_.getFiller()
+                if not isinstance(filler, OWLClass):
+                    raise TypeError(not_el_error_msg)
+                
+                rel_id = self.dataset.object_property_to_id[rel]
+                filler_id = self.dataset.class_to_id[filler]
+                gci_name = "gci2"
+                gci_data = data_point_to_tensor([sub_id, rel_id, filler_id])
+                
+        elif isinstance(sub, OWLObjectSomeValuesFrom):
+            rel = sub.getProperty()
+            filler = sub.getFiller()
+            if not isinstance(filler, OWLClass):
+                raise TypeError(not_el_error_msg)
+            if not isinstance(super_, OWLClass):
+                raise TypeError(not_el_error_msg)
+
+            rel_id = self.dataset.object_property_to_id[rel]
+            filler_id = self.dataset.class_to_id[filler]
+            super_id = self.dataset.class_to_id[super_]
+            if super_.isOWLNothing():
+                if self.extended:
+                    gci_name = "gci3_bot"
+                else:
+                    gci_name = "gci3"
+            else:
+                gci_name = "gci3"
+            
+            gci_data = data_point_to_tensor([rel_id, filler_id, super_id])
+            
+        elif isinstance(sub, OWLObjectIntersectionOf):
+            operands = sub.getOperandsAsList()
+            if len(operands) != 2:
+                raise TypeError(not_el_error_msg)
+            left, right = tuple(operands)
+            if not isinstance(left, OWLClass):
+                raise TypeError(not_el_error_msg)
+            if not isinstance(right, OWLClass):
+                raise TypeError(not_el_error_msg)
+            if not isinstance(super_, OWLClass):
+                raise TypeError(not_el_error_msg)
+
+            left_id = self.dataset.class_to_id[left]
+            right_id = self.dataset.class_to_id[right]
+            super_id = self.dataset.class_to_id[super_]
+
+            if super_.isOWLNothing():
+                if self.extended:
+                    gci_name = "gci1_bot"
+                else:
+                    gci_name = "gci1"
+            else:
+                gci_name = "gci1"
+
+            gci_data = data_point_to_tensor([left_id, right_id, super_id])
+            
+        else:
+            raise TypeError("This axiom does not belong to EL.")
+
+        
+        score = self.module(gci_data, gci_name)
+        return score