+++ mowl/models/__init__.py
@@ -1,0 +1,2 @@
+from mowl.models.elembeddings.model import ELEmbeddings
+from mowl.models.elboxembeddings.model import ELBoxEmbeddings