+++ mowl/nn/__init__.py
@@ -1,1 +1,4 @@
-from .elmodule import ELModule
+from .el.elmodule import ELModule
+from .el.elembeddings.module import ELEmModule
+from .el.elboxembeddings.module import ELBoxModule
+from .el.boxsquaredel.module import BoxSquaredELModule