+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [BoxSquaredEL](https://arxiv.org/abs/2301.11118) module added to `mowl.nn`
 
-## Changed
+### Changed
 - Bug fixed in corpus generation methods. Issue #36.
+
+### Deprecated
+
 
 ## [0.1.1]